Update and rename Spotify Background Image AI to Spotify Background Image AI + Read User Input Login
One of our goals was to create a pleasing image to the eye. Such examples include movement in text, image file parsing, and more. By creating a user login, we can make it more attractive to the users if it would pop out on the page rather than it just be there as is. Another end goal was to read user input and let the user know whether the login info would be accepted into the system and direct them to the following page.